refactor: model entitlement refresh as a mediator command
Move the hub fetch+store into a RefreshEntitlements command (SystemInternal) so
it rides the mediator — command metrics, an event_log audit row per successful
refresh, and readiness for distributed dispatch. EntitlementSyncService is now a
thin fail-open wrapper that binds the mediator scope and dispatches it (mirroring
TenantSnapshotSyncService -> BuildTenantSnapshot). Failures propagate to the
mediator (outcome=failure, no audit row) and the wrapper keeps last-known-good;
not-registered is a no-op. Lengthen the backstop poll to 6h since the revision
header carries freshness.
