+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Added
 
-- **Hub entitlements gate managed-services features (epistola-hub client `0.5.0`).** The hub now hands each installation its configured entitlement set, and the suite uses it to enable/disable features proactively instead of only reacting to a `PERMISSION_DENIED` at call time. A new `EntitlementStore` persists the last-known-good set in `app_metadata` (key `support.entitlements`); `EntitlementSyncService` pulls it via `EpistolaHubClient.getEntitlements()` right after registration and `EntitlementRefreshScheduler` refreshes it hourly (`epistola.support.entitlements.refresh-ms`, default 1h). On any hub failure the previous set is kept (fail-open / last-known-good), so a transient outage never disables a feature. `SupportEntitlementService` resolves a feature for a tenant — a tenant-scoped entry overrides an installation-wide one, and an explicit DENY beats ALLOW — and implements a new core SPI `FeatureEntitlementGate`. A new `ResolveAvailableFeatures` query composes the local toggle with the entitlement: a gated feature (`support-backups`, `support-upgrading`, `support-feedback`) is available only when toggled on **and** entitled. The Backups/Upgrading/Feedback nav contributors, the feedback footer FAB, and the Backups/Upgrading schedulers now consult `ResolveAvailableFeatures`, so gated features are hidden and skipped up front; the reactive `HubEntitlementDeniedException` handling stays as defense-in-depth. **OSS-safe**: with `epistola.support.enabled=false` no gate bean exists and availability is toggle-only, exactly as before. The **Support → Overview** page gained an **Entitlements** section listing the configured entries (feature, scope, allow/deny, expiry) with the last-fetched time — customers see only what's configured for them — plus a **Refresh** button to pull changes on demand. Beyond the hourly backstop poll, the suite refreshes near-instantly on change: the hub stamps an entitlements revision on every response, and `EntitlementRevisionTrigger` (wired to the hub client's `addEntitlementsRevisionListener`) does a debounced async refresh whenever the observed revision exceeds the stored one.
+- **Hub entitlements gate managed-services features (epistola-hub client `0.5.0`).** The hub now hands each installation its configured entitlement set, and the suite uses it to enable/disable features proactively instead of only reacting to a `PERMISSION_DENIED` at call time. A new `EntitlementStore` persists the last-known-good set in `app_metadata` (key `support.entitlements`). The refresh is a mediator command (`RefreshEntitlements`, `SystemInternal`) so it gets command metrics, an `event_log` audit row per successful refresh, and is ready for distributed dispatch; `EntitlementSyncService` is a thin fail-open wrapper that binds the mediator scope and dispatches it (mirroring how `TenantSnapshotSyncService` wraps `BuildTenantSnapshot`). It runs right after registration and on a long backstop poll (`EntitlementRefreshScheduler`, `epistola.support.entitlements.refresh-ms`, default 6h — long because the revision header carries freshness). On any hub failure the previous set is kept (fail-open / last-known-good), so a transient outage never disables a feature and failed refreshes record `outcome=failure` without an audit row. `SupportEntitlementService` resolves a feature for a tenant — a tenant-scoped entry overrides an installation-wide one, and an explicit DENY beats ALLOW — and implements a new core SPI `FeatureEntitlementGate`. A new `ResolveAvailableFeatures` query composes the local toggle with the entitlement: a gated feature (`support-backups`, `support-upgrading`, `support-feedback`) is available only when toggled on **and** entitled. The Backups/Upgrading/Feedback nav contributors, the feedback footer FAB, and the Backups/Upgrading schedulers now consult `ResolveAvailableFeatures`, so gated features are hidden and skipped up front; the reactive `HubEntitlementDeniedException` handling stays as defense-in-depth. **OSS-safe**: with `epistola.support.enabled=false` no gate bean exists and availability is toggle-only, exactly as before. The **Support → Overview** page gained an **Entitlements** section listing the configured entries (feature, scope, allow/deny, expiry) with the last-fetched time — customers see only what's configured for them — plus a **Refresh** button to pull changes on demand. Beyond the backstop poll, the suite refreshes near-instantly on change: the hub stamps an entitlements revision on every response, and `EntitlementRevisionTrigger` (wired to the hub client's `addEntitlementsRevisionListener`) dispatches a debounced refresh whenever the observed revision exceeds the stored one.
 - **Navigation is now module-contributed (`NavContributor` SPI).** The app shell no longer hardcodes the nav menu. A small SPI in `epistola-web` (`app.epistola.suite.htmx.nav`: `NavContributor`/`NavGroup`/`NavItem`/`NavContext`) lets any module declare its own dropdown groups and items, and `NavMenuAggregator` (also in `epistola-web`) injects every contributor, evaluates them per request, drops empty groups, and derives the active section from the request path (replacing the hardcoded `resolveActiveSection` table in `ShellModelInterceptor`). Visibility lives in the contributor — it emits an item only when visible, deciding from a permission predicate and/or its own feature-toggle lookup — so the `feedbackEnabled`/`backupsEnabled`/`upgradingEnabled` shell model attributes are gone and the nav template just iterates `navGroups` (desktop + mobile). The host app ships a `CoreNavContributor` for Authoring/Resources/Operations/Settings; the **base `epistola-support` module now owns the Support group and the Support → Overview page** (moved out of `apps/epistola`), and each feature module (`epistola-support-feedback`/`-backups`/`-upgrading`) ships its own contributor adding its toggle-gated item. One behavior change: a dropdown with no visible items (e.g. Operations for a user with none of its permissions) now hides entirely instead of rendering empty. This is the first concrete UI-contribution extension point described in the commercial-tier architecture.
 - **Footer chrome is module-contributed too (`FooterContributor` SPI).** A sibling SPI in `epistola-web` (`app.epistola.suite.htmx.footer`) lets modules inject Thymeleaf fragments into the shell footer; `FooterFragmentResolver` collects them per request (same `UiRequestContext` as nav). The feedback floating-action-button (and console-capture script), previously hardcoded in **two** host templates — the shell footer (behind a `feedbackEnabled` flag) and the standalone editor page (unconditionally) — now lives once in `epistola-support-feedback` (`templates/feedback/footer-fab.html` + `FeedbackFooterContributor`, gated on `support-feedback`). `footerFragments` is resolved for any tenant view, so both the shell and the editor render contributed footer chrome; the editor FAB now also respects the toggle. The host footer/editor no longer reference any feature, removing the last feedback coupling from `apps/epistola`.
 - **Feature-toggle reads now go through CQRS queries, cached per request.** Rendering and background schedulers stopped injecting `FeatureToggleService` directly and read through a new `ResolveFeatureToggles` query — the `SystemInternal` (auth-bypassing) sibling of the permission-gated `GetFeatureToggles` (which still backs the admin Features page) — so feature visibility resolves for any signed-in user, not just managers. Both delegate to `FeatureToggleService.resolveAll`, which now memoizes per HTTP request via a `ScopedValue` cache bound by `FeatureToggleCacheFilter` (the same idiom as `SecurityContext`/`MediatorContext`); a whole page render — nav contributors, footer FAB, editor flags — issues a single toggle query per tenant instead of one per check. Background schedulers reading outside a request scope are uncached (one query per tenant, as before).
